* Development 15 | Development 1 *
The next update. Lots of new GUI components and some changes and
features on Android. Just a note I am going to change the way
the update cycle works. A few experimental builds = 1 development
builds, a few development builds = 1 beta build and so on

* Added GUISlider
* Added GUIRenderableSlider
* Added GUIImageSlider
* Added GUISliderBuilder
* Added SliderTest
* Added ScrollEvent
* Added GUIScrollBar
* Added GUIImageScrollBar
* Added GUIRenderableScrollBar
* Added GUIScrollBarBuilder
* Added ScrollBarTest
* Added Rectangle
* Added TileMap2D
* Added OpenGLBufferUtils
* Added OpenGLShaderUtils
* Added Shader
* Added FontPack
* Added FontStore
* Added OpenGLUtils
* Added GUIWindow
* Added GUIImageWindow
* Added GUIRendeableWindow
* Added WindowSizeChangedEvent
* Added WindowListener
* Added GUIMessageDialogBox
* Added GUIConfirmDialogBox
* Added GUIInputDialogBox
* Added DialogBoxTest
* Added GUIDropDownMenu
* Added GUIRenderableDropDownMenu
* Added GUIImageDropDownMenu
* Added GUIDropDownMenuTest
* Added GUIDropDownList
* Added GUIImageDropDownList
* Added GUIRenderableDropDownList
* Added GUIDropDownListTest
* Added GameValues

* Added the package org.simplecorporation.myengine.core.game2d.tilemap

* Added the method: movePanel() in GUIPanel
* Added the method: setPanelPosition() in GUIPanel
* Added the method onScroll() to InputListener
* Added the method setupRemoveAlpha() in OpenGLSetupUtils
* Added the abstract method resetEffect() in Effect
* Added the variable maxNumberOfCycles in EffectColourFade
* Added the variable inFolder to read() method in FileUtils

* Added the variable: selected to GUIComponent

* Added scrolling events


* Created package org.simplecorporation.myengine.core.gui.slider
* Created the package org.simplecorporation.myengine.core.rectangle
* Created the package org.simplecorporation.myengine.core.opengl.shader
* Created the package org.simplecorporation.myengine.utils.opengl.shader
* Created the package org.simplecorporation.myengine.core.gui.window
* Created the package org.simplecorporation.myengine.core.event
* Created the package
org.simplecorporation.myengine.core.gui.dropdownlist

* Created the method 'tick()' in EngineLoop

* Changed variable in GUI from the type GUIPanel to GUIComponent
* Changed method in GUIButton isSelected() to isButtonSelected()
* Changed the method name update() to updatePanel() in GUIPanel
* Changed the method name render() to renderPanel() in GUIPanel
* Changed Game to such a way that you should be able to extend it and
use it on android (Only need to extend AndroidActivity then give an
instance of Game) (NOTE: Don't call createGame())
* Changed the parameter: width from the type integer to double in
GUITextDisplayArea
* Changed the way Entity2D is updated. The velocity is multiplied by the
current delta
  then added to the entity's position and the same for the rotation
velocity. This means that the velocity may be different from before ,
however the speed of the computer the game is running on will not effect
the time it takes an entity to move a certain distance.

* Moved Event from org.simplecorporation.myengine.core.input.event to
org.simplecorporation.myengine.core.event
* Moved all game update methods to tick() in EngineLoop instead of
start()
* Moved the 'text' variable in GUIRenderableButton and GUIImageButton to
GUIButton

* Renamed Point2D to Vector2D
* Renamed package org.simplecorporation.myengine.core.game2d.point to
org.simplecorporation.myengine.core.game2d.vector

* Removed everything to do with the old scripting engine (Making new
one)
* Removed AndroidGame
* Removed duplicate method from GUIBuilder

* Fixed bug: Current position of mouse not set when not using OpenGL and
mouse clicked
* Fixed big: If a component in a panel equaled the panel's x, y, width,
or height it didn't update or render
* Fixed bug: If the panel x + height was less than a button it didn't
show (Should of been x + width)
* Fixed bug: IndexOutOfBoundsException when writing an empty file
* Fixed bug: When changing screen size, and using OpenGL, the OpenGL
view port wasn't resized
* Fixed bug: Problems when changing from fullscreen to not fullscreen
when using OpenGL
* Fixed bug: The width of GUILabel was never set
* Fixed bug: The following settings caused a problem
(IllegalStateException):
		Android: false
		Window.Title: Elements
		Window.Fullscreen: true
		Window.Border: true
		Window.Size.Width: 1366.0
		Window.Size.Height: 768.0
		Video.MaxFPS: 60
		Video.VSync: true
		Video.OpenGL: true
		Video.AntiAliasing: true
		Audio.SoundEffectVolume: 100
		Audio.MusicVolument: 100
* Found bug: GUICheckBox doesn't work on Android


* NOTE: GUIPanel now extends GUIComponent
* NOTE: GUI now takes GUIComponent's instead of GUIPanel's
* NOTE: If you wish to move a GUIPanel with components in it use the
method setPanelPosition() and movePanel()
* NOTE: GUI now holds GUIPanels instead of GUIComponents
* NOTE: GUIPanel no longer descends from GUIComponent
* NOTE: It is now possible to read a file from within the same .jar file
* NOTE: You can call methods in OpenGLSetupUtils and OpenGLUtils without
causing error when not using OpenGL
* NOTE: Now using LWJGL V2.9.1
* NOTE: The AndroidGameThread now uses the methods in EngineLoop to
update and
render the game. This means that the FPS and DELTA values are now
measured and the DEBUGGING setting now affects the output of debug info
onto the screen when using MyEngine on Android

+++ MyEngine-Master/src/org/simplecorporation/myengine/core/gui/textdisplayarea/GUITextDisplayArea.java
@@ -23,7 +23,7 @@ public class GUITextDisplayArea extends GUIComponent {
 	/* The font */
 	private GUIFont font;
 	
-	public GUITextDisplayArea(String name , LinkedList<String> textToDisplay , GUIFont font , int width) {
+	public GUITextDisplayArea(String name , LinkedList<String> textToDisplay , GUIFont font , double width) {
 		//Call the super constructor
 		super(name);
 		//Assign the variables
